+++ crates/lib/src/core/parser/grammar/delimited.rs
@@ -25,7 +25,7 @@ pub struct Delimited {
     delimiter: Arc<dyn Matchable>,
     pub(crate) min_delimiters: Option<usize>,
     optional: bool,
-    cache_key: u32,
+    cache_key: u64,
 }
 
 impl Delimited {
@@ -252,7 +252,7 @@ impl Matchable for Delimited {
         Ok(MatchResult { matched_segments, unmatched_segments })
     }
 
-    fn cache_key(&self) -> u32 {
+    fn cache_key(&self) -> u64 {
         self.cache_key
     }
 }
